src/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-autoslug==1.9.8
django-nested-admin==3.3.3
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==8.1.0
psycopg2==2.8.6
PyJWT==2.0.1
python-monkey-business==1.0.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
